+++ backend/src/message/conversation.service.ts
@@ -5,15 +5,12 @@ import { ActiveConversationModel } from "./models/active_conversation.model";
 import { MessageModel } from "./models/message.model";
 import { Op } from "sequelize";
 import { UserModel } from "../user/models/user.model";
-import { UserConversationModel } from "./models/users_conversation.model";
 
 @Injectable()
 export class ConversationService {
   constructor(
     @InjectModel(ConversationModel)
     private conversationRepository: typeof ConversationModel,
-    @InjectModel(UserConversationModel)
-    private userConversationRepository: typeof UserConversationModel,
     @InjectModel(ActiveConversationModel)
     private activeConversationRepository: typeof ActiveConversationModel,
     @InjectModel(MessageModel)
@@ -21,7 +18,7 @@ export class ConversationService {
   ) {}
 
   async getConversations(userId: number) {
-    const conversations = await this.userConversationRepository.findAll({
+    const conversations = await this.conversationRepository.findAll({
       where: {
         [Op.or]: [{ firstUser: userId }, { secondUser: userId }],
       },
@@ -42,7 +39,7 @@ export class ConversationService {
   }
 
   async getConversationWithUsers(firstUser: number, secondUser: number) {
-    return await this.userConversationRepository.findOne({
+    return await this.conversationRepository.findOne({
       where: {
         [Op.and]: [
           { firstUser: firstUser || secondUser },
@@ -54,8 +51,8 @@ export class ConversationService {
   }
 
   async joinConversation(socketId: string, conversationId: number) {
-    const conversation = this.userConversationRepository.findOne({
-      where: { conversationId },
+    const conversation = this.conversationRepository.findOne({
+      where: { id: conversationId },
     });
     if (!conversation) {
       throw new HttpException("No conversation For this user", HttpStatus.NOT_FOUND);
@@ -106,16 +103,11 @@ export class ConversationService {
       return conversation;
     }
 
-    const newConversation = await this.conversationRepository.create();
-
-    await newConversation.save();
-
-    const newUserConversation = await this.userConversationRepository.create({
+    const newConversation = await this.conversationRepository.create({
       firstUser,
       secondUser,
-      conversationId: newConversation.id,
     });
-    await newUserConversation.save();
+    await newConversation.save();
   }
 
   async leaveConversation(socketId: string) {