+++ backend/src/message/message.module.ts
@@ -6,17 +6,11 @@ import { MessageModel } from "./models/message.model";
 import { ActiveConversationModel } from "./models/active_conversation.model";
 import { ConversationModel } from "./models/conversation.model";
 import { MessageGateway } from "./message.gateway";
-import { UserConversationModel } from "./models/users_conversation.model";
 
 @Module({
   imports: [
     AuthModule,
-    SequelizeModule.forFeature([
-      MessageModel,
-      ConversationModel,
-      ActiveConversationModel,
-      UserConversationModel,
-    ]),
+    SequelizeModule.forFeature([MessageModel, ConversationModel, ActiveConversationModel]),
   ],
   providers: [ConversationService, MessageGateway],
   exports: [ConversationService],