+++ backend/src/message/models/conversation.model.ts
@@ -1,12 +1,31 @@
-import { Column, DataType, HasMany, Model, Table } from "sequelize-typescript";
+import {
+  Column,
+  DataType,
+  ForeignKey,
+  HasMany,
+  Model,
+  Table,
+  BelongsTo,
+} from "sequelize-typescript";
 import { ApiProperty } from "@nestjs/swagger";
 import { MessageModel } from "./message.model";
+import { UserModel } from "../../user/models/user.model";
 
 @Table({ tableName: "conversations", createdAt: false, updatedAt: false })
 export class ConversationModel extends Model<ConversationModel> {
   @ApiProperty({ example: "4", description: "Conversation's id" })
   @Column({ type: DataType.INTEGER, unique: true, autoIncrement: true, primaryKey: true })
   id: number;
+
+  @ApiProperty({ example: "23", description: "First user's id" })
+  @Column({ type: DataType.INTEGER, allowNull: false })
+  @BelongsTo(() => UserModel, "id")
+  firstUser: number;
+
+  @ApiProperty({ example: "93", description: "Second user's id" })
+  @Column({ type: DataType.INTEGER, allowNull: false })
+  @BelongsTo(() => UserModel, "id")
+  secondUser: number;
 
   @ApiProperty({ example: [MessageModel], description: "Messages for current conversation" })
   @HasMany(() => MessageModel)